Feat: Add authorization on notes endpoint

+++ src/api/notes/handler.js
@@ -15,7 +15,12 @@ class NotesHandler {
 
     const { title = 'untitle', body, tags } = request.payload;
 
-    const noteId = await this._service.addNote({ title, body, tags });
+    const { id: userId } = request.auth.credentials;
+
+    const noteId = await this._service.addNote({
+      title, body, tags, owner: userId,
+    });
+
     const response = h.response({
       status: 'success',
       message: 'Catatan berhasil ditambahkan',
@@ -28,8 +33,10 @@ class NotesHandler {
     return response;
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: userId } = request.auth.credentials;
+
+    const notes = await this._service.getNotes(userId);
     return {
       status: 'success',
       data: {
@@ -40,6 +47,10 @@ class NotesHandler {
 
   async getNoteByIdHandler(request, h) {
     const { id } = request.params;
+
+    const { id: userId } = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, userId);
+
     const note = await this._service.getNoteById(id);
 
     return h.response({
@@ -51,9 +62,13 @@ class NotesHandler {
   }
 
   async putNoteByIdHandler(request, h) {
+    const { id } = request.params;
+
+    const { id: userId } = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, userId);
+
     this._validator.validateNotePayload(request.payload);
 
-    const { id } = request.params;
     const { title, body, tags } = request.payload;
 
     await this._service.updateNoteById(id, { title, body, tags });
@@ -66,6 +81,10 @@ class NotesHandler {
 
   async deleteNoteByIdHandler(request, h) {
     const { id } = request.params;
+
+    const { id: userId } = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, userId);
+
     await this._service.deleteNoteById(id);
 
     return h.response({
